Bugfixes
- Fixed the intel doors
- Rifle was missing a bracket
- ammoCounter still used all the old variables
- healthcabinet used old variables
- revolver used new variables but only actually set the old ones

All weapons reloaded faster than they could fire, so I set the
reload buffers to refire rate as a temp measure.

Also, blade was a bit of a mess - user 12/13 and the ammo counter
treated it like the minigun, while user 1 treated it like a clip
weapon. I'm lazy and you wanted a rework, so I took the easier
minigun option.