Added festive sprites and hat metagame mechanic
There are now two extra sets of sprites (hats) for characters instead of just one: The Party set (ending in SP) and the Xmas set (ending in SX). The Xmas set was just copied from last year's special Xmas version. Players will start out with normal sprites, but can acquire different ones in several ways: First, there is a chance to get a hat for scoring (capping intel or CP). Second, if you kill someone who has a hat there's a chance that you can steal it. The decision to award and retract hats is made by the server, sent to the those clients that understand the protocol extension and then persisted by the clients in the gg2.ini, so that hats are retained.