Nerfed diagonal rocketjumping plus balance tweaks
While trying to figure out why my rocketjumping nerf
wasn't working, I ended up doing some of the other
balance tweaks, too. I did a test with findude and
sabot but feel nervous as if I missed something. I
changed the bubble sprite because I needed one that
had an even resolution. We can put it back and change
the bubble size by a pixel if it's a problem.