Fix projectiles "coming out" backwardser in 60FPS
Bullets were coming out behind where they should have been coming out
because of the way that movement compensation works. Actually, it's not
because of compensation directly, but because the 30fps movement happened
after the bullets were spawned, before collisions etc, because they're
spawed in a begin step, and then they're newer instances than the one that
ran its begin step, so their begin step ends up running in the same frame.
So the bullets should always move a 30fps movement on their first frame
because that's the practical.

I think this is a better solution than moving bullets in the functions
that spawn them while making it so that they don't run their begin step
the first time.