Fix #25: Taunting with a 60fps client on a 30fps server is likely to cause desync
The desync would make the affected player taunt on the server but not the clients. This fix doesn't entirely eliminate the possibility of taunt desync, but it does make it far less likely to occur. In order to cause visible problems now, a player would have to push the taunt key a really short time, *and* the server would have to receive a second keybyte update in the same step. Still, eventually it would be nice if taunting was synchronized.