Fix numerical text fields in menus not showing their value after editing
item_value[] should (apparently) always be a string, at least it's stringified when initializing a numerical edit item. However, after editing such an item the new value was stored in item_value[] as a real. This apparently never caused any problems because draw_text has no problem with drawing reals, but the addition of the sanitizeNewlines()-call broke this.