Fix ArenaCP capTime scaling
In Arena, the capture time limit used to be changed based on the number (rather, the fraction) of
players still alive on the enemy team. Since we didn't want the capture percentage to jump
around when a player dies, there was a scaling mechanism that adjusted the "capping" time
when the capTime (the limit) changed. However, that adjustment was performed *after* the
capturing for the current frame was already added, meaning that it counted the capping in the
current frame towards the capTime of the previous frame. Not normally noticable, but in some
circumstances related to locking and no players being in one team, this could lead to
instant captures upon CP unlock.

Since we're trying to simulate a situation where the capture percentage is continuous and
only the rate of change is adapted, I reversed the logic so that we now calculate a
cappingPower that scales the rate of change for "capping". Note that, in order to preserve
the old behaviour, "uncapping" is also scaled by the cappingPower of the team that currently
has the point partially captured, not by the cappingPower of the team who tries to uncap.
Whether this makes sense is a decision for a later day, since this is only supposed to be
a bugfix commit.