Balance: Limit the capping advantage in Arena to 4x the normal capping speed (suggested by footpop)
The smaller the fraction of living players in the enemy team, the lower the capTime (~ the larger the capping power). This commit limits the capping power so that it won't cause excessively fast captures in matches with a large team where only one player is still alive. To be honest I'm not convinced of the usefulness of the entire mechanic...