Turn Bubbleface into a "normal" bubble, update aFirst flag regularly
Fixes Gang-Garrison-2/Gang-Garrison-2#241
The chosen solution is to have a "placeholder" bubble in the actual BubbleS sprite that is never drawn, instead the draw code in SpeechBubbleO specially draws BubbleFaceS for that index. This should cause the least issues with existing code / mods since we still use image_index to control which bubble is active. Inlining BubbleFaceS into BubbleS properly would lead to a lot of blank/wasted texture space because BubbleFaceS is a lot larger.